+++ src/components/header/homeheader.tsx
@@ -6,6 +6,7 @@ import { Users, FolderPen, Image as ImageIcon, Component, Moon, Sun, Settings, S
 import { useUser } from "@clerk/clerk-react"
 import { useState, useEffect } from "react"
 import { useRouter } from "next/navigation"
+import { usePathname } from 'next/navigation';
 import {
   DropdownMenu,
   DropdownMenuContent,
@@ -59,7 +60,7 @@ export default function HomeHeader({activesection}: any) {
   const router = useRouter()
   const { user } = useUser()
   const [isMenuOpen, setIsMenuOpen] = useState(false)
-
+  const pathname = usePathname();
 
   return (
   <header className="w-full z-50 bg-transparent">
@@ -125,7 +126,7 @@ export default function HomeHeader({activesection}: any) {
       <Link href={item.route} key={item.label}>
         <Button
         variant="ghost"
-        className={`w-full justify-start font-semibold py-2 ${router?.pathname === item.route ? 'text-blue-500' : ''}`}
+        className={`w-full justify-start font-semibold py-2 ${pathname === item.route ? 'text-blue-500' : ''}`}
         >
         <item.icon className="mr-2 h-4 w-4" />
         {item.label}
